[#36] Extract icon select sheet
