+++ src/screens/DayForm/DayFormScreen.tsx
@@ -1,35 +1,28 @@
 import { yupResolver } from "@hookform/resolvers/yup";
 import { useNavigation, useRoute } from "@react-navigation/native";
 import { type NativeStackScreenProps } from "@react-navigation/native-stack";
-import React, { useEffect, useRef, useState } from "react";
+import React, { useRef } from "react";
 import { useController, useForm } from "react-hook-form";
 import { type TFunction, useTranslation } from "react-i18next";
-import { type TextInput as RNPTextInput, ScrollView, StyleSheet, View } from "react-native";
-import {
-  Button,
-  IconButton,
-  SegmentedButtons,
-  Text,
-  TextInput as TextFieldRNP,
-} from "react-native-paper";
+import { type TextInput as RNPTextInput, ScrollView, StyleSheet } from "react-native";
+import { Button, IconButton, SegmentedButtons } from "react-native-paper";
 import { v4 as uuidv4 } from "uuid";
 import * as yup from "yup";
 
-import { BottomSheet, type BottomSheetRef } from "@components/dialogs";
+import { type BottomSheetRef, IconSelectSheet } from "@components/dialogs";
 import { Checkbox, DateTimeInput, TextInput } from "@components/form";
 import { DayIcon } from "@components/icons";
 import { AppBar, Page, Stack } from "@components/layout";
-import { useAppTheme, useDayActions, useDebouncedValue, useMounted } from "@hooks";
+import { useAppTheme, useDayActions, useMounted } from "@hooks";
 import { type MaterialCommunityIcons } from "@typings/app.types";
 import { type DayUnit } from "@typings/day.types";
-import { dayIcons } from "@utilities/icons.util";
 import { type RootRouterParams } from "src/AppRouter";
 
 type DayFormScreenRouteProps = NativeStackScreenProps<RootRouterParams, "DayFormScreen">;
 
 interface IFormData {
   date: string;
-  icon?: keyof MaterialCommunityIcons;
+  icon: keyof MaterialCommunityIcons | null;
   title: string;
   repeats: boolean;
   unit: DayUnit;
@@ -67,29 +60,11 @@ const DayFormScreen = () => {
   const dateRef = useRef<RNPTextInput | null>(null);
 
   const iconSheetRef = useRef<BottomSheetRef | null>(null);
-  const [iconSearch, setIconSearch] = useState("");
-
-  // Limit how often icons are filtered via debouncing
-  const filteredSearch = useDebouncedValue(iconSearch, 100);
-  const [filteredIcons, setFilteredIcons] = useState<(keyof MaterialCommunityIcons)[]>([]);
-
-  // TODO: Figure out how to improve performance (mainly rendering issues???)
-  useEffect(() => {
-    // TODO: Remove spaces/hyphens when searching (in icon name tags)
-    const simpleSearch = filteredSearch.toLowerCase().replace(/[^a-z]gi/, "");
-    const icons = dayIcons.reduce((accum, icon) => {
-      if (!simpleSearch) return [...accum, icon.name];
-      const matches =
-        icon.name.includes(simpleSearch) || icon.tags.some((tag) => tag.includes(simpleSearch));
-      return matches ? [...accum, icon.name] : accum;
-    }, [] as (keyof MaterialCommunityIcons)[]);
-    setFilteredIcons(icons);
-  }, [filteredSearch]);
 
   const form = useForm<IFormData>({
     defaultValues: {
       date: editingDay?.date ?? "",
-      icon: editingDay?.icon ?? undefined,
+      icon: editingDay?.icon ?? null,
       title: editingDay?.title ?? "",
       repeats: editingDay?.repeats ?? false,
       unit: editingDay?.unit ?? "day",
@@ -142,12 +117,12 @@ const DayFormScreen = () => {
   };
 
   const onIconSelectChoose = (icon: keyof MaterialCommunityIcons) => {
-    form.setValue("icon", iconValue !== icon ? icon : undefined);
+    form.setValue("icon", iconValue !== icon ? icon : null);
     iconSheetRef.current?.close();
   };
 
   const onIconClear = () => {
-    form.setValue("icon", undefined);
+    form.setValue("icon", null);
   };
 
   const { onDayAdd, onDayEdit } = useDayActions({
@@ -267,43 +242,8 @@ const DayFormScreen = () => {
           </Button>
         </Stack>
       </ScrollView>
-      <BottomSheet ref={iconSheetRef} dismissable title={"Select Icon"}>
-        <TextFieldRNP
-          autoCapitalize="none"
-          dense
-          placeholder="Search"
-          left={<TextFieldRNP.Icon icon="magnify" />}
-          right={
-            !!iconSearch && <TextFieldRNP.Icon icon="close" onPress={() => setIconSearch("")} />
-          }
-          value={iconSearch}
-          onChangeText={setIconSearch}
-        />
-        {/* TODO: Capture height on mount and maintain even when filtering (prevents flicker) */}
-        <ScrollView
-          contentContainerStyle={{
-            flexWrap: "wrap",
-            flexDirection: "row",
-            gap: 4,
-            paddingVertical: 4,
-          }}
-          style={{ marginTop: 8, height: 200 }}
-        >
-          {filteredIcons.map((icon) => (
-            <View key={icon}>
-              <IconButton
-                containerColor={icon === iconValue ? colors.primaryContainer : undefined}
-                iconColor={icon === iconValue ? colors.onPrimaryContainer : undefined}
-                icon={icon}
-                size={24}
-                onPress={() => onIconSelectChoose(icon)}
-                style={{ margin: 0 }}
-              />
-            </View>
-          ))}
-          {!filteredIcons.length && <Text style={{ paddingHorizontal: 16 }}>No icons found</Text>}
-        </ScrollView>
-      </BottomSheet>
+
+      <IconSelectSheet ref={iconSheetRef} value={iconValue} onSelect={onIconSelectChoose} />
     </Page>
   );
 };
